Move MetricsProvider classes in new file and fix var names
Moving MetricsProvider classes and subclasses (and MetricsCollector) will allow to share the file with jobwatcher/nodewatcher which will use it when adding custom metrics for SGE/Torque. Variable names and comments added.
