+++ tests/slurm_plugin/test_clustermgtd.py
@@ -79,8 +79,7 @@ class TestClustermgtdConfig:
                     "disable_scheduled_event_health_check": False,
                     "disable_all_health_checks": False,
                     "health_check_timeout": 180,
-                    # Disable pushing cluster metrics to cloudwatch
-                    "disable_push_metrics_to_cloudwatch": False,
+                    "disable_cloudwatch_metrics": False,
                 },
             ),
             (
@@ -112,8 +111,7 @@ class TestClustermgtdConfig:
                     "disable_scheduled_event_health_check": True,
                     "disable_all_health_checks": False,
                     "health_check_timeout": 10,
-                    # Disable pushing cluster metrics to cloudwatch
-                    "disable_push_metrics_to_cloudwatch": True,
+                    "disable_cloudwatch_metrics": True,
                 },
             ),
             (
