+++ tests/slurm_plugin/test_clustermgtd/TestClustermgtdConfig/test_config_parsing/all_options.conf
@@ -24,4 +24,4 @@ master_hostname = master-hostname
 hosted_zone = hosted-zone
 dns_domain = dns.domain
 use_private_hostname = false
-disable_push_metrics_to_cloudwatch = True
+disable_cloudwatch_metrics = True
